@@ -1,5 +1,7 @@
 module tideland.dev/go/actor
 
-go 1.19
+go 1.24
 
-require tideland.dev/go/audit v0.7.0
+require tideland.dev/go/asserts v0.2.1
+
+require golang.org/x/exp v0.0.0-20240325151524-a685a6edb6d8 // indirect
