@@ -1,6 +1,6 @@
 // Tideland Go Actor - Unit Tests
 //
-// Copyright (C) 2019-2023 Frank Mueller / Tideland / Oldenburg / Germany
+// Copyright (C) 2019-2025 Frank Mueller / Tideland / Oldenburg / Germany
 //
 // All rights reserved. Use of this source code is governed
 // by the new BSD license.
@@ -18,7 +18,7 @@ import (
 	"testing"
 	"time"
 
-	"tideland.dev/go/audit/asserts"
+	"tideland.dev/go/asserts/verify"
 
 	"tideland.dev/go/actor"
 )
@@ -29,107 +29,101 @@ import (
 
 // TestPureOK verifies the starting and stopping an Actor.
 func TestPureOK(t *testing.T) {
-	assert := asserts.NewTesting(t, asserts.FailStop)
 	finalized := make(chan struct{})
 	act, err := actor.Go(actor.WithFinalizer(func(err error) error {
 		defer close(finalized)
 		return err
 	}))
-	assert.OK(err)
-	assert.NotNil(act)
+	verify.NoError(t, err)
+	verify.NotNil(t, act)
 
 	act.Stop()
 
 	<-finalized
 
-	assert.NoError(act.Err())
+	verify.NoError(t, act.Err())
 }
 
 // TestPureDoubleStop verifies stopping an Actor twice.
 func TestPureDoubleStop(t *testing.T) {
-	assert := asserts.NewTesting(t, asserts.FailStop)
 	finalized := make(chan struct{})
 	act, err := actor.Go(actor.WithFinalizer(func(err error) error {
 		defer close(finalized)
 		return err
 	}))
-	assert.OK(err)
-	assert.NotNil(act)
+	verify.NoError(t, err)
+	verify.NotNil(t,act)
 
 	act.Stop()
 	act.Stop()
 
 	<-finalized
 
-	assert.NoError(act.Err())
+	verify.NoError(t, act.Err())
 }
 
 // TestPureError verifies starting and stopping an Actor.
 // Returning the stop error.
 func TestPureError(t *testing.T) {
-	assert := asserts.NewTesting(t, asserts.FailStop)
 	finalized := make(chan struct{})
 	act, err := actor.Go(actor.WithFinalizer(func(err error) error {
 		defer close(finalized)
 		return errors.New("damn")
 	}))
-	assert.OK(err)
-	assert.NotNil(act)
+	verify.NoError(t, err)
+	verify.NotNil(t, act)
 
 	act.Stop()
 
 	<-finalized
 
-	assert.ErrorMatch(act.Err(), "damn")
+	verify.ErrorMatch(t, act.Err(), "damn")
 }
 
 // TestContext verifies starting and stopping an Actor
 // with an external context.
 func TestContext(t *testing.T) {
-	assert := asserts.NewTesting(t, asserts.FailStop)
 	ctx, cancel := context.WithCancel(context.Background())
 	act, err := actor.Go(actor.WithContext(ctx))
-	assert.OK(err)
-	assert.NotNil(act)
+	verify.NoError(t, err)
+	verify.NotNil(t, act)
 
 	cancel()
-	assert.NoError(act.Err())
+	verify.NoError(t, act.Err())
 }
 
 // TestSync verifies synchronous calls.
 func TestSync(t *testing.T) {
-	assert := asserts.NewTesting(t, asserts.FailStop)
 	finalized := make(chan struct{})
 	act, err := actor.Go(actor.WithFinalizer(func(err error) error {
 		defer close(finalized)
 		return err
 	}))
-	assert.OK(err)
+	verify.NoError(t, err)
 
 	counter := 0
 
 	for i := 0; i < 5; i++ {
-		assert.OK(act.DoSync(func() {
+		verify.NoError(t, act.DoSync(func() {
 			counter++
 		}))
 	}
 
-	assert.Equal(counter, 5)
+	verify.Equal(t, counter, 5)
 
 	act.Stop()
 
 	<-finalized
 
-	assert.ErrorMatch(act.DoSync(func() {
+	verify.ErrorMatch(t, act.DoSync(func() {
 		counter++
 	}), "actor is done")
 }
 
 // TestTimeout verifies timout error of a synchronous Action.
 func TestTimeout(t *testing.T) {
-	assert := asserts.NewTesting(t, asserts.FailStop)
 	act, err := actor.Go()
-	assert.OK(err)
+	verify.NoError(t, err)
 
 	// Scenario: Timeout is shorter than needed time, so error
 	// is returned.
@@ -137,13 +131,13 @@ func TestTimeout(t *testing.T) {
 	err = act.DoSyncWithContext(ctx, func() {
 		time.Sleep(25 * time.Millisecond)
 	})
-	assert.NoError(err)
+	verify.NoError(t, err)
 	cancel()
 	ctx, cancel = context.WithTimeout(context.Background(), 50*time.Millisecond)
 	err = act.DoSyncWithContext(ctx, func() {
 		time.Sleep(100 * time.Millisecond)
 	})
-	assert.ErrorMatch(err, "action.*context deadline exceeded.*")
+	verify.ErrorMatch(t, err, "action.*context deadline exceeded.*")
 	cancel()
 
 	time.Sleep(150 * time.Millisecond)
@@ -153,21 +147,20 @@ func TestTimeout(t *testing.T) {
 // TestWithTimeoutContext verifies timout error of a synchronous Action
 // when the Actor is configured with a context timeout.
 func TestWithTimeoutContext(t *testing.T) {
-	assert := asserts.NewTesting(t, asserts.FailStop)
 	ctx, cancel := context.WithTimeout(context.Background(), 50*time.Millisecond)
 	act, err := actor.Go(actor.WithContext(ctx))
-	assert.OK(err)
+	verify.NoError(t, err)
 
 	// Scenario: Configured timeout is shorter than needed
 	// time, so error is returned.
 	err = act.DoSync(func() {
 		time.Sleep(10 * time.Millisecond)
 	})
-	assert.NoError(err)
+	verify.NoError(t, err)
 	err = act.DoSync(func() {
 		time.Sleep(100 * time.Millisecond)
 	})
-	assert.ErrorMatch(err, "actor.*context deadline exceeded.*")
+	verify.ErrorMatch(t, err, "actor.*context deadline exceeded.*")
 
 	act.Stop()
 	cancel()
@@ -175,9 +168,8 @@ func TestWithTimeoutContext(t *testing.T) {
 
 // TestAsyncWithQueueCap tests running multiple calls asynchronously.
 func TestAsyncWithQueueCap(t *testing.T) {
-	assert := asserts.NewTesting(t, asserts.FailStop)
 	act, err := actor.Go(actor.WithQueueCap(128))
-	assert.OK(err)
+	verify.NoError(t, err)
 
 	sigs := make(chan struct{}, 1)
 	done := make(chan struct{}, 1)
@@ -198,7 +190,7 @@ func TestAsyncWithQueueCap(t *testing.T) {
 	// Now start asynchrounous calls.
 	now := time.Now()
 	for i := 0; i < 128; i++ {
-		assert.OK(act.DoAsync(func() {
+		verify.NoError(t,act.DoAsync(func() {
 			time.Sleep(2 * time.Millisecond)
 			sigs <- struct{}{}
 		}))
@@ -209,14 +201,13 @@ func TestAsyncWithQueueCap(t *testing.T) {
 	<-done
 	duration := time.Since(now)
 
-	assert.OK((duration - 250*time.Millisecond) > enqueued)
+	verify.True(t, (duration-250*time.Millisecond) > enqueued)
 
 	act.Stop()
 }
 
 // TestRecovererOK tests successful handling of panic recoveries.
-func TestNotifierOK(t *testing.T) {
-	assert := asserts.NewTesting(t, asserts.FailStop)
+func TestRecovererOK(t *testing.T) {
 	counter := 0
 	recovered := false
 	done := make(chan struct{})
@@ -226,7 +217,7 @@ func TestNotifierOK(t *testing.T) {
 		return nil
 	}
 	act, err := actor.Go(actor.WithRecoverer(recoverer))
-	assert.OK(err)
+	verify.NoError(t, err)
 
 	act.DoSync(func() {
 		counter++
@@ -234,19 +225,18 @@ func TestNotifierOK(t *testing.T) {
 		fmt.Printf("%v", counter/(counter-1))
 	})
 	<-done
-	assert.True(recovered)
+	verify.True(t, recovered)
 	err = act.DoSync(func() {
 		counter++
 	})
-	assert.OK(err)
-	assert.Equal(counter, 2)
+	verify.NoError(t, err)
+	verify.Equal(t, counter, 2)
 
 	act.Stop()
 }
 
 // TestRecovererFail tests failing handling of panic recoveries.
-func TestNotifierFail(t *testing.T) {
-	assert := asserts.NewTesting(t, asserts.FailStop)
+func TestRecovererFail(t *testing.T) {
 	counter := 0
 	recovered := false
 	done := make(chan struct{})
@@ -256,7 +246,7 @@ func TestNotifierFail(t *testing.T) {
 		return fmt.Errorf("ouch: %v", reason)
 	}
 	act, err := actor.Go(actor.WithRecoverer(recoverer))
-	assert.OK(err)
+	verify.NoError(t, err)
 
 	act.DoSync(func() {
 		counter++
@@ -264,10 +254,10 @@ func TestNotifierFail(t *testing.T) {
 		fmt.Printf("%v", counter/(counter-1))
 	})
 	<-done
-	assert.True(recovered)
+	verify.True(t, recovered)
 
-	assert.True(act.IsDone())
-	assert.ErrorMatch(act.Err(), "ouch:.*")
+	verify.True(t, act.IsDone())
+	verify.ErrorMatch(t, act.Err(), "ouch:.*")
 }
 
 // EOF
